Add files via upload
This project is aimed at improving my Excel skills through practice done by using this  project to analyze the Bike Sales dataset so as to derive  a data-driven customer insights and as well provide a data driven answers to specific questions that beg for urgent answers. The analytical recommendation will subsequently help the company to achieve their desire goals.